backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-tavily==0.2.11
tavily-python==0.7.10
groq==0.31.0
python-dotenv==1.1.1
pydantic==2.11.7
requests==2.32.4
